[Gradle] Fix 'android.kotlinOptions' DSL produces compilation errors in Gradle generated accessors
On the one hand, Gradle generated accessors do not carry over
'@Deprecated' annotations from the plugin types, so users are not aware
about deprecation. On the other hand, Gradle does not suppress the
deprecation error in the generated accessor. Which prevents users from
importing the project due to accessor compilation error.

For now, this fix lifts the deprecation level to warning. Added warnings
 on the DSL properties should lead users to stop using this DSL.

^KT-77288 Verification Pending