[K/JS] Fix an issue with overriding message property from Throwable
In case of using the default `Throwable` constructor, the custom error classes may set the `message` field
which was not created by the JavaScript `Error` class (if `undefined` was provided as a message parameter).
To escape such a situation, we always provide `null` to the `message` parameter if nothing was provided.

^KT-77271 Fixed


Merge-request: KT-MR-21558
Merged-by: Artem Kobzar <Artem.Kobzar@jetbrains.com>