[JS] Explicitly pass through hasErrors flag in CLI pipeline artifacts
If there were errors on the first compilation stage, we need to
set the `contains_error_code` flag in the KLIB manifest.
Previously, we'd just take that flag from the diagnostic reporter
right before serialization, assuming that the same diagnostic reporter
was used in the whole pipeline up to serialization.

However, this is not the case in tests: each test facade creates
a new diagnostic reporter and passes it to the corresponding phase,
thus the `contains_error_code` may not be set.

Fix it by checking the diagnostic reporter fort errors at the end of
each pipeline phase and propagating the `hasError` value down to
other phases.

This is required to make this test
 `compiler/testData/ir/irText/expressions/badBreakContinue.kt` pass
 after switching the irText test to the CLI-based
 test infra (KT-77193)