[FIR] Implement `PLACEHOLDER_PROJECTION_IN_QUALIFIER_RESOLUTION`
Initially, I wanted to route the diagnostic through `FirTypeCandidateCollector`,
but found we also allow `FirResolvedQualifier`s with placeholder arguments
in some other scenarios, not related to callable references. One of them
was `Generic<_>.Nested()` - a case that even has its own deprecation
cycle, so we can't just break it and call it a bugfix :( See:
`ProhibitGenericQualifiersOnConstructorCalls`, targeting Kotlin 2.2.

^KT-76914 Fixed
^KT-73527