[LL] disable hidden deprecation check for source Java classes
This check is too unpredictable due to Java resolution,
so there are no guaranties that there are no contract violations.
The "feature" itself is not widely used, so it should be fine
to disable it for sources.
The compiler for now still sticks to the previous behavior,
but in the context of KT-77114 in the compiler it should be disabled
as well

^KT-76789 Fixed
^KT-77114