Scripts: write backend outputs in the legacy CLI pipeline
This fixes an IDE test `KotlinProjectStructureTest.test compiled script
class always belongs to KaLibraryModule`.

Before 5207fbe35d, outputs were written as a part of a supposed
optimization, which however did not work as an optimization because the
outputs were written only at the end. Now we must write them manually
instead. Note that in the new phased CLI, this is done in the phase
`JvmWriteOutputsPhase`.