IrSymbolDeserializer: Simplify 'deserializeIrSymbolToDeclare()'
This function is intended to be used only during deserialization of
IR declarations. So, when a declaration is deserialized,
a signature/symbol pair needs to be deserialized first. And this
pair can always be deserialized in the scope of the _current_ file --
without involving the IR linker (i.e. without
`deserializePublicSymbol()` call).

^KT-65837