[K/N] Enforce the native thread state for skiko's _nFlushAndSubmit
`org_jetbrains_skia_DirectContext__1nFlushAndSubmit` is a function from
Skiko written in C++ and called from Kotlin through a
`@SymbolName external fun`.

As discovered in KT-75895, this function might block and wait for
another Kotlin thread. Therefore, it violates the contract of
`SymbolName`. Switching the thread state to "native" is required when
calling such a function, which `SymbolName` doesn't do.

This commit adds a hack to the compiler: when calling a function with
such a symbol name, enforce thread state switching.

The hack is configurable: it is added in the form of a binary option,
`-Xbinary=forceNativeThreadStateForFunctions=fun1;fun2`, with
`org_jetbrains_skia_DirectContext__1nFlushAndSubmit` as the default
value.
It can also be disabled with
`-Xbinary=forceNativeThreadStateForFunctions=`.

Specifying a value for `forceNativeThreadStateForFunctions` disabled
compiler caches (since callsites to listed functions might be in the
caches).

^KT-77489 Fixed