[CFG] Sort flow variables in the CFG graph render output
Flow variables are stored in unordered persistent maps. If we render
them as is, we get a graph we cannot store in the test data, or compare
against.

Before, 'allVariablesForDebug' were sorted by a name, but it was only a
partial solution. In particular, 'renderHtmlLike()' for synthetic
variables can return the same values, while aliased variables can be
different. Implications also come unsorted.

In this commit, output values are compared by a name (to preserve the
ordering in non-problematic cases), and then by the whole text (so
ordering becomes stable in all cases).

This solves flakiness in partial body analysis tests, when the resulting
CFG is compared against CFG of a non-partial body analysis.

^KT-72357