[CFG] Substitute references to FirElements in CFG snapshots
State keepers in the LL API replace all unresolved statements with the
new ones, as the previous iteration of analysis could be unsuccessful
(e.g., it could be interrupted by a PCE). However, in some cases, the
CFG holds references to an unresolved part (see the 'MiniCfgBuilder' and
'FirLocalVariableAssignmentAnalyzer.buildInfoForRoot').

This commit injects an element substitutor to 'createSnapshot()',
allowing the LL API to seamlessly adapt references to the new ones.

^KT-72357