[K/JS] Fix `FileNotFoundException` exception in incremental caches writing
It seems like the root cause of all the flaky `FileNotFoundException` is a case,
when some of the caches were invalidated (so their directory was deleted). So, we
tried to write a cached `*.js.map` file into deleted by the previous compilation directory.
I haven't figured out how to write such a test case because even the reported behavior is flaky.

^KT-73267 Fixed


Merge-request: KT-MR-21676
Merged-by: Artem Kobzar <Artem.Kobzar@jetbrains.com>