K2: report additional NEW_INFERENCE_ERROR for special non-inferred variables
This commit considers as special the type variables which aren't
based on some type parameter. In this case we cannot prove that
CANNOT_INFER_*_TYPE diagnostic will be reported, so we have to report
NEW_INFERENCE_ERROR to avoid possible diagnostic loss. See also
the comment added to coneDiagnosticToFirDiagnostic.kt

Related to KT-77361