[AA][FIR] Fix missing and unresolved `partiallyResolvedTypeRef` in local type refs
Fix consists of two parts:
- Ensure proper level of resolve of `partiallyResolvedTypeRef`
in `FirAbstractBodyResolveTransformerDispatcher`.
This typeRef can be unresolved at this moment when the transformer
is executed in the IDE mode (under Analysis API).
- Properly copy `partiallyResolvedTypeRef`
in `withReplacedSourceAndType` function. Without it, this information
does not survive until Analysis API can use it to properly resolve
qualified types with errors in them.

This change allows properly resolving qualified types with errors
with Analysis API (KT-71794, KT-73216), and also prevents exceptions
when doing so (KT-77516).

Update `partiallyUnresolvedTypeQualifier` test to also cover type refs
in local positions and lambda parameters positions

^KT-71794 Fixed
^KT-77516 Fixed
^KT-73216 Fixed
^KTIJ-32210 Fixed