[Analysis API] Split the implementation of `KaTypeInformationProvider.getHasFlexibleNullability` for K1/K2
The problem is that dynamic types are a special case of flexible types, even though in AA we don't consider them as such.
Generally speaking, a dynamic type is a flexible type with bounds of `Nothing` and `Any?`.
However, it's important to note that a dynamic type doesn't always have a flexible nullability, every platform can assign custom bounds to dynamic types.
So the implementation had to be moved to K1/K2 modules to be able to acquire additional info on bounds of dynamic type in K2 (in K1 all dynamic types don't hold any additional information).

^KT-71101