IR Deserialization: Topo-sort of IR modules in K/JS backend
In upcoming commits we are going to get rid of the KLIB resolver.
However, with the removal of KLIB resolver we are also getting rid of
sorting of IR modules using the reverse topological order (RTO).
This can have negative consequences on the Kotlin/JS backend until we
implement OSIP-695 initiative: Make lowerings independent of
the IR modules order.

In this commit, we introduce a temporary component
`IrModuleDependencyTracker` that will be used for tracking cross-module
dependencies and then sorting the IR modules in the backend.

^KT-77242
^KT-65837