IR Deserialization: Enable topo-sort of IR modules in K/JS backend
In preceding commits we are getting rid of the KLIB resolver. However,
with the removal of KLIB resolver we are also getting rid of sorting of
IR modules using the reverse topological order (RTO). This can have
negative consequences on the Kotlin/JS backend until we implement
OSIP-695 initiative: Make lowerings independent of the IR modules order.

In this commit, we enable RTO for IR modules in Kotlin/JS backends
via the previously introduced `IrModuleDependencyTracker` component.

^KT-77242 Fixed
^KT-65837