[FIR] Add readiness `WITH_COMPLEX_DEPENDENCY_AND_NO_CONCRETE_CONSTRAINTS`
It's not enough to relax it just to `WITH_COMPLEX_DEPENDENCY`, as we'd
still pick `T` for fixation first, and we'd still try to fix it to
`Comparable<Int>`, leading to contradictions.

Before this change, we had...
1. `T` with `WITH_COMPLEX_DEPENDENCY_BUT_PROPER_EQUALITY_CONSTRAINT`
2. `S` with `WITH_COMPLEX_DEPENDENCY`
...and after we now have...
1. `T` with `WITH_COMPLEX_DEPENDENCY_AND_NO_CONCRETE_CONSTRAINTS`
2. `S` with `WITH_COMPLEX_DEPENDENCY`
Now we first fix `S := Long` and then get `T := Long`.
Also, see: `e892a10`.

^KT-67335 Fixed

# Conflicts:
#	compiler/testData/diagnostics/testsWithStdLib/greater.inference.md
#	compiler/testData/diagnostics/testsWithStdLib/greater.inference.mmd