[FIR] Add readiness `WITH_COMPLEX_DEPENDENCY_AND_PRIMITIVE_CONSTRAINT`
It's not enough to relax it just to `WITH_COMPLEX_DEPENDENCY`, as we'd
still pick `T` for fixation first, and we'd still try to fix it to
`Comparable<Int>`, leading to contradictions.

Before this change, we had...
1. `T` with `WITH_COMPLEX_DEPENDENCY_BUT_PROPER_EQUALITY_CONSTRAINT`
2. `S` with `WITH_COMPLEX_DEPENDENCY`
...and after we now have...
1. `T` with `WITH_COMPLEX_DEPENDENCY`
2. `S` with `WITH_COMPLEX_DEPENDENCY_AND_PROPER_NON_ILT`
Now we first fix `S := Long` and then get `T := Long`.
Also, see: `e892a10`.

^KT-67335 Fixed