[Tests] Mark constraints edges with colors
When the graph gets a bit wide and a bunch of nodes get clamped together,
some nodes may overlay nearby edges, so it becomes hard to tell if an
edge originates from a specific node or a node behind it. This change
adds colors to nearby nodes into different hues to help detangle the
mess.