[CFG] Track in-place versus non-in-place assignments in UnusedChecker
To correctly report if a variable assignment is read or not, each
variable read must be tracked according to its in-place relationship
with the corresponding variable write. Variable writes which happen
"later" in the code may still be visible to "earlier" reads if they are
not within the same in-place scope. More information about the
implementation details has been added to the checker documentation.

^KT-75338 Fixed