[IR] Always remap type parameters of extracted locals from inline lambda
Previously we thought that lambdas couldn't introduce new type
parameters, so the remapping wasn't required. This statement breaks in
the following example
```
inline fun <T> inlineGenericTestFunction(f: () -> T) = f()

val updateStatus = inlineGenericTestFunction {
    fun <F> bangbang(flag: F) = flag!! // new type parameter `F`
}
```

Here the local function inside lambda has a type parameter. It is
used as a resulting type of `!!` expression, and so it must be
remapped when the local function is extracted. Otherwise, we end up
with a type parameter without a parent.

#KT-77102 Fixed