[JS] Fix getting names of local delegated properties
The test would fail with "Fail 2: property2" because invoking
the property2 delegate triggered a side effect due to
a global function being mutated.
Namely, in `getLocalDelegateReference` a contextless lambda
was created and later passed to `getPropertyCallableRef`,
which mutated it. However, because the lambda was contextless,
it was transformed into a global function, so the global
function was mutated instead, which caused unexpected side effects.

^KT-78073 Fixed

(cherry picked from commit fe7dda91c58f5e86e293f51114c9082a356cffca)