[Inline] Replace originalFunction with a dedicated attribute
In fact, the only use case of originalFunction (a.k.a. attributeOwnerId)
here is for a case of a special inline function copy used in 1-st stage
inlining. In all other cases callee.originalFunction == callee.
To indicate that we now use a dedicated attribute instead of
attributeOwnerId, which is rather JVM specific, and in general very
vague.

#KT-76512 Fixed
KT-76763