[Inline] Don't follow IrRichCallableReference.attributeOwnerId
In all circumstances functionArgument.attributeOwnerId will also be
an IrRichCallableReference, but possibly a different instance of it.

For KLIB backends there is no difference, as only the type of a node is
used.
For JVM backend, once it migrates to rich references, it will be more
consistent with the current way (the
`functionArgument is IrFunctionExpression` branch above), where
attributeOwnerId is not used.

KT-76763
KT-78175