[Inline] Don't follow IrFunction.attributeOwnerId in inline()
In all circumstances callee.attributeOwnerId == callee.

For it to not be equal, the callee must have been a local inline
function, and those are not allowed in the language. Even if they were,
it would be problematic to follow such function,
at least in KLIB backends, where it would have a local signature.

KT-76763