[K/N] Fix HostExecutor timeout handling
On timeout `HostExecutor` used to forcibly
destroy the process. This does not play nice
with `xcrun simctl spawn`: forcibly terminating
it does not terminate its child process (the
one we're actually testing), so it may hang
until the simulator is shut down.

This commit makes `HostExecutor` wait 1 minute
to allow the process to terminate gracefully.

^KT-78043