[CLI] Set proper delimiter for `-Xfragment-dependencies` arguments
Previously it was `,` but it's incorrect to use it, as comma is a valid
character for a path. Default path separator cannot be used too, as
on Unix systems it's `:`, which is used in the syntax of arguments
themselves. So eventually it was decided to set the delimiter to `""`
(no delimiter).

^KT-78129 Fixed