[FIR] Use nullability of recursive upper bounds in inference
This fixes a case where a type argument with a recursive upper bound
could not be inferred from a nullable expected type.

`canBeNull` is properly implemented for `ConeTypeVariableType`.
This required changing two usages (through `isNullableType`).

In `ConstraintIncorporator.isPotentialUsefulNullabilityConstraint`,
we implicitly relied on the fact that a TV T was not
considered nullable. This behavior is preserved using a newly
introduced parameter `considerTypeVariableBounds`.

In
`FirControlFlowStatementsResolveTransformer.transformElvisExpression`,
we implicitly relied on `canBeNull` returning false for TVs to never
trigger the logic for making elvis type not-null or flexible inside
call arguments.
In fact, it would be incorrect to do so because the RHS could be an
unresolved lambda. But also, by changing the nullability of an argument,
a contradiction in the CS can be provoked.
See compiler/testData/diagnostics/tests/controlStructures/lambdasInExclExclAndElvis.kt.

#KT-76826 Fixed