Stdlib: setup compilation of JVM stdlib with return-value-checker=full
Other targets are compiled with CHECKER mode + suppression for now
to allow usage of @IgnorableReturnValue (which requires RVC flag to be on).

It is also required to suppress ACTUAL_ANNOTATION_NOT_MATCH_EXPECT because of
annotations on things like MutableList which have their own implementations
on e.g. wasm. This suppression will be removed in subsequent commits.

#KT-71195
#KT-78462
#KT-78463