[Analysis API] Clean up and refactor the super type approximation infrastructure in `KaTypeProvider`
The previous API `approximateToSuperPublicDenotable` got deprecated, as it had a misleading naming (it has nothing to do with `public` declarations) and `approximateLocalTypes` parameter had shady meaning and wasn't actually consistent with the provided documentation.
It was replaced with `approximateToSuperDenotable`, which always performs local types approximation, however, has a way to choose whether local types can be approximated to denotable local types or not.

This is done by introducing a new filter for local types in `TypeApproximatorConfiguration`.

^KT-68387