[Maven, Daemon] Instead of using -Xjava-source-roots, pass .java files along Kotlin ones
It's important to pass .java files to the incremental compiler runner so that it could keep track of changes in them, and that should be consistent between in-process and daemon executions.
Additionally, there's no sense in -Xjava-source-roots, but only additional overhead in traversing the roots to find the Java sources by the compiler once again even though we already found them anyway.
^KT-77587 Verification Pending