Stdlib, JVM: Add a continuation for tail-call suspend functions
Currently, tail-call functions are missing from async stack
in the debugger. This is because to build the async stack,
the debugger uses continuation chain and when there is
no continuation, the function is invisible for the debugger.
To combat this, add a function in stdlib, which is called
from tail-call suspend functions and by default, just returns
the continuation, which is supplied to it. However, when
debugger is attached, the debugger is expected to replace
the body of the function with one, which wraps the
continuation with fictitious one, which does not store any
variables or even debug metadata, and only holds stack
trace element, so the tail-call functions will appear in
async stack.

 #KT-74051 Fixed