[IR] Don't pad argument list with nulls when changing callee
When reassigning IrMemberAccessExpression.symbol to a function that has
more parameters than before, new nulls were appended to the arguments
list. This was only added to maintain a correct old<->new parameter API
bridge. It is quite fragile and potentially unexpected mechanism, it is
better to be removed and left for manual patches.

KT-76934