[K/N][runtime] Handle thread states in termination more gracefully
`terminateWithUnhandledException` calls `consoleFlush` from the runnable
thread state, but this call is potentially blocking and also not
listed among CallsChecker exclusions.

Therefore, when `terminateWithUnhandledException` is called with the
thread state checker enabled, the latter reports a problem.

This commit fixes the incorrect state by correctly switching the thread
state for `consoleFlush`.
It also ignores the thread state checker there: we are more interested
in the exception than in the incorrect thread state reports at this
stage.