[K/N][tests] Help Gradle foojay resolver use ARM64 JDK on ARM64 Mac
Starting from version 0.9.0, Gradle foojay resolver may use x86_64 on
ARM64 Mac:

https://plugins.gradle.org/plugin/org.gradle.toolchains.foojay-resolver-convention/0.9.0

This doesn't work well for the K/N compiler tests:
the tests in the repo use JDK 1.8 by default, and it seems that there is
no x86_64 version of it in the toolchain repositories.

So, when running on an ARM64 Mac, the build uses an ARM64 JDK (and thus
produces the ARM64 version of the compiler), and the tests run on an
x86_64 JDK (so the compiler fails to load the built JNI libraries).

This commit workarounds this by making the affected test tasks use
JDK 11.0, for which an ARM64 version is available in the repos.

See KTI-2421.