[K/N] Add IR attribute for an original name of local function
instead using attributeOwnerId. In pursue of removing attributeOwnerId
in klib backends (KT-78175).

It is used to obtain a correct name of a function for reflection, in
case it was lifted and renamed by LocalDeclarationsLowering.

Note that there is a bug with klib inliner mode (KT-78537), because
attributeOwnerId is not serialized to klib. This commit does not fix it,
as IR attribute is neither, but at least makes it easier to track.

KT-78175
KT-78537