K2/inference: introduce precise mode of lower constraints simplification
The idea of this commit is to simplify Foo? <: T! not to Foo? <: T,
which is dangerous and can provoke information lost,
but to (Foo!! .. Foo?) <: T instead which is safer.
After KTLC-284 this change is necessary, otherwise we get
type mismatches in obviously wrong situations.
Foo <: T! is in turn simplified to (Foo!! .. Foo) <: T.

#KT-78621 Fixed