[IR][inliner] Removed an old workaround
The workaround was for per-file lowering invocation (the whole pipeline including the inliner).
This is not needed anymore since now the inliner and all earlier lowerings are invoked per-module.
Moreover, this workaround may run the prefix (before the inliner) twice on the same file which is
not desired (since some of the lowerings there are not idempotent).