[FIR] Set resolvedToCompanionObject when resolving through a typealias
Clean up resolved qualifier logic in the process.

Set the explicit parent in some cases where it was previously missed.

The logic for unsetting resolvedToCompanionObject is removed from
buildResolvedQualifierForClass because it's now called from tower resolution and
we mustn't modify FIR at this stage because the candidate won't necessarily be
selected.
Instead, it's called from FirCallResolver after the final candidate is selected.