[stdlib] Use XOR for computing BigInt hash code in JS
Once we make Longs BigInt-backed (KT-57128), this implementation
will be consistent with how hashCode is implemented for Longs on
other targets.

Namely, in other targets it's just `((value ushr 32) xor value).toInt()`
which is equivalent to this implementation as long as the width of
a BigInt value is at most 64 bits, which will be always true for
BigInt-backed Longs.