[FIR] Extract inlinability checks from InlineFunctionBodyContext
Checks inside inline functions are split into two groups:

InlineFunctionBodyContext is responsible for visibility, access and
recursion checks.

InlinableParameterContext is responsible for inlinability checks.

This is required so that we can disable the former without affecting
the latter in a following commit.

For variable assignments we don't need to call the inlinability checks
so we skip straight to the inline body checks.