[Gradle] Fix testFixtures compilation classpath may miss declared dependencies
Gradle test-fixtures plugin creates the own set of configurations which
are not connected to the Kotlin compilation created for test fixtures.
This leads to a different set of dependencies depending on which
configuration they were declared.

Fix ensures both compile and runtime classpaths are synced between the
Java and Kotlin set of configurations.

^KT-77466 Verification Pending