[FIR2IR] Check intersection components against approximated expected type
When the expected type is an intersection type with an upper bound for
approximation, one of the components could be a subtype of the upper
bound without being a subtype of the intersection type.

This fixes a K1/K2 difference in one box test which makes the test pass
on Wasm.

#KT-78849 Fixed