[LL] add consistency check to stub-based resolution tests
During stub-based resolution we might jump to the original file due to
KT-78949, so some elements might have a higher phase in the output.
For instance, an implicit return type from a property might be calculated
and returned from the original file. In this case, the stub-based
property itself won't be marked as resolved as expected, but some other
element, which uses the return type, might have a higher phase as
it would miss the exception from the property

^KT-77097
^KT-78949