[PowerAssert] Do not extract lambda arguments to inline functions
Power-Assert regularly extracts local variables even if they are not
used in the error message to maintain the correct order of operations
of the original code. Power-Assert was mistakenly extracting all
lambdas, including those to inline functions. Instead, if a function
parameter is going to be inlined, do not extract it to a local variable.
This doesn't change the order of operations as the lambda is going to be
inlined.

^KT-77626 Fixed