[IR] Add unsigned functions to interpreter function map
The generated source file can now also handle function calls for
UByte, UShort, UInt and ULong.

The extension functions on primitives to convert to unsigned integers
couldn't be added yet because the current compiler would already try
to interpret such calls and eventually fail when they aren't found in
the builtins map.
So we have to do this in two stages, first add it to the map
(which this commit does) and then later add the annotation to the
functions  once a bootstrapped compiler is available.

KT-51065