[K/N] Remove kotlin-native.jar from the dist
The K/N distribution contains two jars with compiler code:
* konan/lib/kotlin-native.jar
* konan/lib/kotlin-native-compiler-embeddable.jar

The former is basically a legacy version of the latter. It is used only
by KGP when `kotlin.native.useEmbeddableCompilerJar=false` is set.

Since this flag is no longer supported (KT-51301), it is safe to
remove the jar from the distribution. This commit does that.
As a nice side effect, it also noticeably reduces the size of the
distribution.

^KT-77783 Fixed