Scripting: implement and use component for compatibility with dataframe
Add a temporary FirSession component that allow to skip tower data
cleanup when resolving some initializers in scripts. It is needed to
allow the temporary fix of dataframe plugin compatibility with the
scripts, since they (the scripts) are currently used by the Kotlin
Notebook implementation in IntelliJ. See #KT-78724 for details.
These changes should be reverted as soon as KN will switch to the
REPL snippet implementation in the IDE. See #KT-79107