[CLI] Read annotations manually for PsiPackage in single file mode
A custom implementation of PsiPackage is used when returning information
about Java sources. The default implementation of annotations only works
when reading from binaries or when Java sources are specified as a
directory. Annotations need to be read manually from the PsiFile if Java
sources are specified to the compiler as individual files.

^KT-77729 Fixed