Implement partial KMP dependency resolution diagnostic
Walk over the dependency graph similarly to how we do this in
metadata transformation task, but at configuration time. If we detect
that metadata variant or one of platform variants resolved, but some
other platform variants didn't resolve, report this as a lenient error.

The lenient error will be rendered as a red node in IDE, but will not
block compilations.

^KT-69790