FIR: Do not account for hasMustUseReturnValue while checking main() function status
It is allowed to have multiple `public fun main(): Unit` functions in the same package in different files.
To achieve that, we check that the function signature is equal to the one we expect from main().
However, certain flags such as hasMustUseReturnValue and isSuspend should not affect this check,
because main() function is allowed but not obliged to have them.

#KT-79085 Fixed