[FIR] Only report FUNCTION_CALL_EXPECTED when candidate has correct receiver
When we don't find any candidates for a non-function call, we try
searching candidates as if it was a function call and then report
`FUNCTION_CALL_EXPECTED`.
However, this is confusing when all found candidates have incorrect
receivers because it suggests that the call would be successful if
was converted to a function call.
After this commit, we only report `FUNCTION_CALL_EXPECTED` if the
found candidates have `applicability >= INAPPLICABLE_WRONG_RECEIVER`
ensuring that if the call was converted to a function call, there is
a way to make it successful.

#KT-77104 Fixed