[K/N][tests] Implement a Gradle init script for filtering tests
When running K/N tests, it might be useful to limit the test run to
certain tests.

We already have filtering by JUnit test tags
(`-Pkotlin.native.tests.tags`), but it doesn't allow easily including
only a single test method, for example.

Gradle has the `--tests` flag for that, but it works only when running a
test task explicitly.
For example, test tasks that are executed as dependencies aren't
covered, which makes this flag not applicable to `:nativeCompilerTest`.

This commit closes that gap by providing a Gradle init script:
`kotlin-native/tools/includeOnlyTests/includeOnlyTests.init.gradle.kts`.

When using this script, specify the project property:
```
./gradlew \
    -I kotlin-native/tools/includeOnlyTests/includeOnlyTests.init.gradle.kts \
    -Pkotlin.includeOnlyTests=pattern1,pattern2,... \
```

The patterns listed there are added to `Test` tasks as
[includeTestsMatching](https://docs.gradle.org/current/kotlin-dsl/gradle/org.gradle.api.tasks.testing/-test-filter/include-tests-matching.html).

This new Gradle init script is not specific to K/N tests or even the
Kotlin project. It can in fact be applied to any Gradle project.