[Lowerings] Improve cross-file f/o function reference classes
Before this commit, if a fun interface is already lowered,
at the point where a reference class for it's implementation is
generated, the current list of declarations was used to build fake
overrides.

This is not correct, as this can lead to double application of
further lowerings to these overrides.

This commit starts storing current set of declarations to attribute,
so that fake overrides could be built correctly.

Change is only applied for non-jvm backends for now, as jvm
handles this second application of lowering correctly.

^KT-78831
^KT-79092