KT-78701 Include not-found enum value in enumValueOf exception message
enumValueOf exception message is supposed to include a value that was not found, but in case of Js and Wasm it was referring to the function's parameter name instead.
That seems to be a bug and this PR aligns (to some extent) exception message's content with other targets.
^KT-78701

Merge-request: KT-MR-22462
Merged-by: Filipp Zhinkin <filipp.zhinkin@jetbrains.com>