[decompiler] implement stub-based decompiler
The commit introduces a new decompiler to replace
the previous one which depends on K1 (descriptions deserialization).

### Work pipeline

The previous one:
1. We have `A.class` file
2. `KotlinClassFileDecompiler` builds a compiled stub via reading metadata.
This stub is used for indexing
3. Someone opened the `A.class` file in the editor
4. `KtClsFile` is created
5. To build the decompiled text for it, `buildDecompiledTextForClassFile`
is used to read K1 descriptors from the metadata again and then render
to text

The new approach has a different `5.` step. Instead of reading the metadata
again, the compiled stub from the step `2.` is used to render
the decompiled text.
This reduces the overhead on the metadata parsing (which was caused by
K1 descriptors deserialization) and makes the decompiler implementation
frontend-agnostic.
Also, it is much easier to pass the information to the decompiler since
now we own the entire pipeline

^KT-74547 Fixed
^KT-79484 Fixed
^KT-79487 Fixed
^KT-60764 Fixed