[K/JS] Fix incremental compilation for overridden external properties
There was an issue with a non-external class overriding an external
property with the incremental compilation.

```kotlin
external interface Foo { val a: Int }
class FooImpl(override val a: Int): Foo
```

In the case when such a module is not marked as dirty (so it will be not
re-compiled), the lowering was failing with an internal compiler
exception. The body of the ` FooImpl ` constructor was
not loaded during the re-compilation.

^KT-79050 Fixed


Merge-request: KT-MR-22573
Merged-by: Artem Kobzar <Artem.Kobzar@jetbrains.com>