FIR: allow to resolve some Java classes with `@kotlin.Metadata`
Namely, Java classes coming from sources, as well as any Java classes
annotated with `@kotlin.Metadata` without the "kind" (JVM name "k")
argument.

This is needed to fix the "unresolved reference" problem in kapt when
generated stubs are used as a dependency when generating stubs for
another module.

Note that there's still a K1/K2 difference in the
differentModulesWithKind.kt test, however it's not a part of the issue,
so K2 behavior there is not changed.

 #KT-79138 Fixed