Fix misc code to get rid of `PLATFORM_CLASS_MAPPED_TO_KOTLIN` diagnostic
It's necessary because the full Kotlin is compiled
by a bootstrap version with enabled `-Werror`.

In some places suppression is used because Java class literals
with primitive types don't have corresponding Kotlin counterparts.
I.e. `java.lang.Integer` is not equivalent to `kotlin.Int`,
but `java.lang.Object` is equivalent to `kotlin.Any`.