[FIR] Avoid unnecessary `MUST_BE_INITIALIZED_OR_BE_ABSTRACT`
Yeah, this condition is super-cryptic and I had great trouble trying
to figure out what it means in K1, so here's a separate commit
highlighting its purpose.

If a property isn't `lateinit`, then it's definitely initialized in the
constructor and is not accessed too early (giving us an error otherwise).
If a property is `lateinit`, then we need to check if it has to be.

^KT-59807