[IR] Only generate static F/Os for members defined in Java
Fake overrides for static members are only generated to overcome an
visibility issue in class inheritance in Java. Because it cannot happen
in Kotlin code, there is no need to generate static F/Os for Kotlin
members. Doing so confused non-JVM backends, as they are not prepared
to handle a static member that overrides something.

#KT-78960 Fixed