[Serialization] Fix deserialization of nested type aliases
All type aliases (top-level and nested) should be cached during deserialization as well as classes.
However, top-level and nested type aliases are deserialized in different ways:

* Info about top-level type aliases is extracted from package parts
* Info about nested type aliases is stored in their parent class protos
  and currently isn't available in `findAndDeserializeTypeAlias`

To make it possible to use the `findAndDeserializeTypeAlias` caching function
for type aliases of any kind,
the `FirNestedTypeAliasDeserializationContext` is introduced.

It allows passing of nested type aliases proto into the caching function
and avoiding useless searching for top-level type aliases when it's useless.

^KT-79519 Fixed