Reorganize access to "customized" features in LanguageVersionSettings
Right now, there is no guaranteed way to get the full set of language
features that have been customized in `LanguageVersionSettings`.
I.e., which features were explicitly specified in the LANGUAGE
directive of a test data file with "plus" and "minus" signs
irrespectively of their state for the specific language version used
in that test.

There are two auxiliary functions that might look like they do exactly
what is required. But in fact they are not:
`getManuallyEnabledLanguageFeatures()` and
`getManuallyDisabledLanguageFeatures()`. These functions return ONLY
meaningful language features. And they filter out any enabled feature
that is already "stable" in the current language version, or any
disabled language feature that is "experimental" in the current LV.

So, we are introducing a new function in `LanguageVersionSettings`
to get the full set of customized features:
`getCustomizedLanguageFeatures()`. Also, the two auxiliary functions
are renamed to `getCustomizedEffectively[En|Dis]abledLanguageFeatures()`
to correctly reflect their meaning.

^KT-71957
^KT-78188