Fix equals methods when comparing primitive range with ComparableRange
In my codebase I added a helper method for working with Comparable<T>
that creates a range using T..T and I was surprised to see that my tests
failed when I was asserting the values against a literal 1L..2L.

This PR fixes that behavior by checking if the other object is of type
ClosedRange and then compares the interface fields directly.  I couldn't
get this to work for Short/Byte because the operator functions for those
types return an IntRange (which cannot easily be coerced to/from a
ComparableRange<Byte>, etc).  Open to suggestions on how/whether this
can be fixed!

Also added a ReadMe to the generators/builtins folder because I couldn't
figure out how to run those generators from the other ReadMe files.