JVM: workaround inline class mapping issue in old backend
The real fix will be removal of `StackValue` and
`StaticTypeMapperForOldBackend`, and it will be done as a part of
KT-71197. This change is a relatively safe workaround for the issue.

The problem is caused by 588361c70d, where
`computeExpandedTypeForInlineClass` started to use more things from
`TypeSystemContext`, namely `builtIns`, to be able to correctly map
generic types based on inline classes over arrays.

 #KT-79547 Fixed