[IR] Don't validate referenced declarations
There are a couple of issues with way it's done currently:
- The same invalid declaration may be reported many times.
- The validation is recomputed many times for the same IR.
- The set of checkers that validates both declared and referenced
declarations is arbitrary, but quite low.
- It requires some unnecessary duplication of code.

Instead, on the first phase of compilation, the code should be validated
only by the module that defines it. On the second phase, we traverse over
the entire IR, so all declarations will be reached anyway.

The problematic case may be lazy IR in JVM backend, because it is
being lowered just a bit, but is not visited by validator. However,
it is not transformed in a way that would be useful to validate with
the removed checkers. If we see would be useful in general, we should
introduce a generic mechanism to run all checkers on lazy IR in the
future.