PL tests: Allow two types of delimiters in *.info files
Formerly, only comma (`,`) was considered as a valid separator in
test setting in an *.info file. Example:

arguments: -Xfoo, -Xbar

Now, the space character is also a valid separator. This will make it
more similar to the CLI experience when arguments are passed just
separated by the space.

arguments: -Xfoo -Xbar