PL tests: Don't use inline funs from stdlib at problematic call sites
Consider, we have a problematic function `fun foo()`. And it is
being called from an inline function from stdlib like this:

run(::foo)

Then the PL engine will generate different error messages for the cases
when stdlib was built with IR inliner enabled and when it was not.

To avoid this ambiguity in PL tests, it's better to avoid calling
problematic functions using inline functions from stdlib. For
the example above this means that we can add a semantically equivalent
non-inline function `fun myRun()` and call `foo()` from there.

fun <R> myRun(block: () -> R): R = block()
myRun(::foo)