[FIR] Add a note on what `getIntersectedLowerType()` does in `LogicSystem`
In theory, we could also introduce a hack that preserves components of
the form `¬(B & B)` specifically: if a lower type is present in every
statement, append it to the result. But it could introduce inconsistencies
in cases like `s: ¬(Base | A) || s: ¬(Base | B) || s: ¬(Derived | C)`:
we'd miss `¬(Base & Base & Derived)`, thus losing the final `¬Derived`.

^KT-76635