K2: introduce deprecation for use-site unrelated missing dependency super-type
Before this commit, we had an error for this case
(see UnrelatedInaccessibleSuperType.kt/.fir.kt) in K1
and green code in K2 till 2.2.20.
Beginning from 2.2.20 we have an error again,
thanks to 56bcef85b6efd5ba3da013a89339e05018613fdf.
In this commit we introduce a deprecation warning for this case
(reported in 2.2.20-2.3.* according to the current plan),
that will become an error in language version 2.4.

#KT-79789 Fixed