[PSI] Introduce KtDeclarationWithReturnType interface
* Implement it in KtCallableDeclaration, KtPropertyAccessor,
  KtBackingField, KtEnumEntry and KtDestructuringDeclaration

* Deprecate `returnTypeReference` property in favor of `typeReference`
  override in KtPropertyAccessor and KtBackingField

Note: From the FIR point of view, a KtDestructuringDeclaration is
a FirProperty whose return type is the type of its initializer. So,
to preserve existing behavior, it too extends KtDeclarationWithReturnType
(also, this case looks important for UAST, see
FirKotlinUastResolveProviderService#getTargetType).

Note: the code of `KtDeclarationWithReturnType` is written in Java
instead of Kotlin to preserve compatibility (see KT-79894 for details).

KT-59857