K/Wasm, Tests: Add missing configuration parameters in env configurator
There configuration parameters are necessary to run KLIB
backward/forward compatibility tests. In these tests the compiler
configuration is used as an input for a custom compiler invocation.

The infrastructure for running the compatibility tests is common
between Kotlin/JS and Kotlin/Wasm. But in Kotlin/JS, which is already
migrated to CLI-based pipelines, there are already all the necessary
parameters set in `JsEnvironmentConfigurator`. So, shall we do in
`WasmEnvironmentConfigurator`.

^KT-78188